fixed s conversion and handle unsigned numbers

+++ cx_conversion.c
@@ -6,7 +6,7 @@
 /*   By: varnaud <varnaud@student.42.us.org>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/04 01:09:56 by varnaud           #+#    #+#             */
-/*   Updated: 2016/11/21 17:32:20 by varnaud          ###   ########.fr       */
+/*   Updated: 2017/01/10 17:03:42 by varnaud          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,5 +16,5 @@
 
 int		cx_conversion(t_flags *flags, va_list *args)
 {
-	return ((int)x_conversion(flags, args));
+	return (x_conversion(flags, args));
 }
